+++ gaios/Receive/Cells/ReceiveAssetCell/ReceiveAssetCell.swift
@@ -5,7 +5,8 @@ class ReceiveAssetCell: UITableViewCell {
     @IBOutlet weak var bg: UIView!
     @IBOutlet weak var imgView: UIImageView!
     @IBOutlet weak var lblAsset: UILabel!
-
+    @IBOutlet weak var iconDisclose: UIImageView!
+    
     class var identifier: String { return String(describing: self) }
 
     var onTap: (() -> Void)?
@@ -20,6 +21,7 @@ class ReceiveAssetCell: UITableViewCell {
         self.lblAsset.text = model.title
         self.imgView?.image = model.icon
         self.onTap = onTap
+        self.iconDisclose.isHidden = onTap == nil
     }
     @IBAction func tap(_ sender: Any) {
         onTap?()