+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+- New home screen asset view navigation
+
 ## [5.0.6] - 2025-07-02
 
 ### Fixed